New test and implementation
Test adds new entry into SyncObject and checks if after one update the copy of GetTourState dictionary is equal to it. Now also implemented. (+6 squashed commit)

Squashed commit:

[50dc291] New Test and Implementation

Added a test to check if time stamp will be updated. After test was failing added comparison of timestamp and change of timestamp in the GetTourState script.

[546eaba] First test in Unity

Inplemented a test to see if the GetTourState starts in the right way and creates a SyncObject with the expected data

[f64c2d5] Temporary changes

[1509b65] Temporary changes

[e83877e] gitignore change, SyncObject erstellen

SyncObject mit Konstruktor und allen ?bergabeparametern. Dictonary Erstellung

[9a30493] .gitignore ?nderungen

Assets hinzugef?gt